refactor(adapters): Refactor to provide generic adapter and specific postgres adapter
